// src/types/api.d.ts
export interface getPlaces_GetMyPlaces_places {
  __typename: "Place";
  id: number;
  name: string;
  address: string;
  isFav: boolean;
}

export interface getPlaces_GetMyPlaces {
  __typename: "GetMyPlacesResponse";
  ok: boolean;
  error: string | null;
  places: (getPlaces_GetMyPlaces_places | null)[] | null;
}

export interface getPlaces {
  GetMyPlaces: getPlaces_GetMyPlaces;
}